Created a new component for member information
